requirements.txt
numpy
pandas
streamlit
python-dotenv
fastapi
mistral-common
langchain_openai
langchain-core
transformers